refactor: moved esbuild plugin to config, moved vitest shared path to config
