+++ .github/workflows/ci.yml
@@ -62,7 +62,6 @@ jobs:
 
   lint:
     runs-on: ubuntu-latest
-    needs: [build]
     steps:
       - uses: actions/checkout@v3
       - uses: ./.github/actions/setup-node
