+++ generator/templates/basic-js/package.json
@@ -35,6 +35,7 @@
     "test": "vitest run"
   },
   "devDependencies": {
+    "@templ/config": "workspace:*",
     "@types/node": "^20.11.19",
     "@vitest/coverage-v8": "^1.3.0",
     "vitest": "^1.3.0"
