+++ generator/templates/basic-js/vitest.config.mjs
@@ -1,7 +1,7 @@
 import { defineProject, mergeConfig } from 'vitest/config'
-import configShared from '../../../.config/vitest.shared.js'
+import { getVitestConfig } from '@templ/config'
 
 export default mergeConfig(
-  configShared,
+  getVitestConfig(),
   defineProject({}),
 )
