+++ package.json
@@ -13,9 +13,7 @@
   "sideEffects": true,
   "workspaces": [
     "apps/**",
-    ".config/**",
     "packages/**",
-    "plugins/**",
     "scripts",
     "generator"
   ],
