+++ packages/api/README.md
@@ -40,12 +40,6 @@ npx nypm i @templ/api@^0.0.1
 # ✨ Auto-detect
 npx nypm i @templ/api@^0.0.1
 
-# ✨ Auto-detect
-npx nypm i @templ/api@^0.0.1
-
-# ✨ Auto-detect
-npx nypm i @templ/api@^0.0.1
-
 # npm
 npm install @templ/api@^0.0.1
