+++ packages/config/package.json
@@ -36,15 +36,18 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
+    "@lukeed/ms": "^2.0.2",
     "c12": "^1.8.0",
     "consola": "^3.2.3",
+    "cpy": "^11.0.0",
+    "esbuild": "^0.20.1",
+    "globby": "^14.0.1",
     "pkg-types": "^1.0.3",
     "scule": "^1.3.0",
     "std-env": "^3.7.0",
     "untyped": "^1.4.2"
   },
   "devDependencies": {
-    "@templ/esbuild-plugin-templ": "workspace:*",
     "@templ/tsconfig": "workspace:*",
     "@types/node": "^20.11.19",
     "@vitest/coverage-v8": "^1.3.0",
@@ -53,5 +56,5 @@
     "typescript": "^5.3.3",
     "vitest": "^1.3.0"
   },
-  "size-limit": "2.5 KB"
+  "size-limit": "5 KB"
 }
