+++ packages/config/src/plugins/esbuild/dts.ts
