+++ packages/config/src/plugins/esbuild/index.ts
