+++ packages/config/src/plugins/index.ts
@@ -1,0 +1,1 @@
+export * from './esbuild'
