--- packages/fflag/README.md
@@ -1,58 +1,0 @@
-# @templ/fflag
-
-TEMPL fflag module.
-
-## `badges`
-
-<!-- automd:badges license provider=shields -->
-
-[![npm version](https://img.shields.io/npm/v/@templ/fflag)](https://npmjs.com/package/@templ/fflag)
-[![npm downloads](https://img.shields.io/npm/dm/@templ/fflag)](https://npmjs.com/package/@templ/fflag)
-[![license](https://img.shields.io/github/license/rjoydip/templ)](https://github.com/rjoydip/templ/blob/main/LICENSE)
-
-<!-- /automd -->
-
-## `pm-x`
-
-<!-- automd:pm-x args=. -->
-
-```sh
-# npm
-npx @templ/fflag@0.0.1 .
-
-# pnpm
-pnpm dlx @templ/fflag@0.0.1 .
-
-# bun
-bunx @templ/fflag@0.0.1 .
-```
-
-<!-- /automd -->
-
-## `pm-install`
-
-<!-- automd:pm-install -->
-
-```sh
-# ✨ Auto-detect
-npx nypm i @templ/fflag@^0.0.1
-
-# ✨ Auto-detect
-npx nypm i @templ/fflag@^0.0.1
-
-# npm
-npm install @templ/fflag@^0.0.1
-
-# yarn
-yarn add @templ/fflag@^0.0.1
-
-# pnpm
-pnpm install @templ/fflag@^0.0.1
-
-# bun
-bun install @templ/fflag@^0.0.1
-```
-
-<!-- /automd -->
-
-[GitHub](https://github.com/rjoydip/templ/tree/main/packages/fflag)
