--- packages/fflag/package.json
@@ -1,53 +1,0 @@
-{
-  "name": "@templ/fflag",
-  "type": "module",
-  "version": "0.0.1",
-  "private": true,
-  "packageManager": "pnpm@8.15.3",
-  "description": "fflag",
-  "license": "MIT",
-  "homepage": "https://github.com/rjoydip/templ/tree/main/packages/fflag#readme",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/rjoydip/templ.git",
-    "directory": "packages/fflag"
-  },
-  "bugs": {
-    "url": "https://github.com/rjoydip/templ/issues"
-  },
-  "exports": {
-    "types": "./dist/index.d.ts",
-    "import": "./dist/index.js",
-    "require": "./dist/index.js"
-  },
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "*.d.ts",
-    "dist"
-  ],
-  "engines": {
-    "node": "^18.8.0 || >=20.6.0",
-    "npm": ">=8"
-  },
-  "scripts": {
-    "build": "tsx esbuild.config.ts",
-    "dev": "tsx esbuild.config.ts --watch",
-    "test": "vitest run",
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "rc9": "^2.1.1"
-  },
-  "devDependencies": {
-    "@templ/config": "workspace:*",
-    "@templ/tsconfig": "workspace:*",
-    "@types/node": "^20.11.19",
-    "@vitest/coverage-v8": "^1.3.0",
-    "esbuild": "^0.20.1",
-    "tsx": "^4.7.1",
-    "typescript": "^5.3.3",
-    "vitest": "^1.3.0"
-  },
-  "size-limit": "1024 KB"
-}
