+++ packages/flags/README.md
@@ -1,0 +1,64 @@
+# @templ/flags
+
+TEMPL flags module.
+
+## `badges`
+
+<!-- automd:badges license provider=shields -->
+
+[![npm version](https://img.shields.io/npm/v/@templ/flags)](https://npmjs.com/package/@templ/flags)
+[![npm downloads](https://img.shields.io/npm/dm/@templ/flags)](https://npmjs.com/package/@templ/flags)
+[![license](https://img.shields.io/github/license/rjoydip/templ)](https://github.com/rjoydip/templ/blob/main/LICENSE)
+
+<!-- /automd -->
+
+## `pm-x`
+
+<!-- automd:pm-x args=. -->
+
+```sh
+# npm
+npx @templ/flags@0.0.1 .
+
+# pnpm
+pnpm dlx @templ/flags@0.0.1 .
+
+# bun
+bunx @templ/flags@0.0.1 .
+```
+
+<!-- /automd -->
+
+## `pm-install`
+
+<!-- automd:pm-install -->
+
+```sh
+# ✨ Auto-detect
+npx nypm i @templ/flags@^0.0.1
+
+# ✨ Auto-detect
+npx nypm i @templ/flags@^0.0.1
+
+# ✨ Auto-detect
+npx nypm i @templ/flags@^0.0.1
+
+# ✨ Auto-detect
+npx nypm i @templ/flags@^0.0.1
+
+# npm
+npm install @templ/flags@^0.0.1
+
+# yarn
+yarn add @templ/flags@^0.0.1
+
+# pnpm
+pnpm install @templ/flags@^0.0.1
+
+# bun
+bun install @templ/flags@^0.0.1
+```
+
+<!-- /automd -->
+
+[GitHub](https://github.com/rjoydip/templ/tree/main/packages/flags)
