+++ packages/flags/esbuild.config.ts
