+++ packages/flags/package.json
@@ -1,16 +1,16 @@
 {
-  "name": "@templ/esbuild-plugin-templ",
+  "name": "@templ/flags",
   "type": "module",
   "version": "0.0.1",
   "private": true,
   "packageManager": "pnpm@8.15.3",
-  "description": "esbuild-plugin-templ",
+  "description": "flags",
   "license": "MIT",
-  "homepage": "https://github.com/rjoydip/templ/tree/main/plugins/esbuild-plugin-templ#readme",
+  "homepage": "https://github.com/rjoydip/templ/tree/main/packages/flags#readme",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/rjoydip/templ.git",
-    "directory": "plugins/esbuild-plugin-templ"
+    "directory": "packages/flags"
   },
   "bugs": {
     "url": "https://github.com/rjoydip/templ/issues"
@@ -31,20 +31,23 @@
     "npm": ">=8"
   },
   "scripts": {
-    "build": "tsx ./esbuild.config.ts",
+    "build": "tsx esbuild.config.ts",
+    "dev": "tsx esbuild.config.ts --watch",
+    "test": "vitest run",
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@lukeed/ms": "^2.0.2",
-    "consola": "^3.2.3",
-    "cpy": "^11.0.0",
-    "esbuild": "^0.20.1",
-    "globby": "^14.0.1",
-    "typescript": "^5.3.3"
+    "rc9": "^2.1.1"
   },
   "devDependencies": {
+    "@templ/config": "workspace:*",
     "@templ/tsconfig": "workspace:*",
     "@types/node": "^20.11.19",
-    "tsx": "^4.7.1"
-  }
+    "@vitest/coverage-v8": "^1.3.0",
+    "esbuild": "^0.20.1",
+    "tsx": "^4.7.1",
+    "typescript": "^5.3.3",
+    "vitest": "^1.3.0"
+  },
+  "size-limit": "1024 KB"
 }
