--- plugins/esbuild-plugin-templ/CHANGELOG.md
@@ -1,7 +1,0 @@
-# @templ/esbuild-plugin-templ
-
-## 0.0.1
-
-### Minor Changes
-
-- minor release
