--- plugins/esbuild-plugin-templ/README.md
@@ -1,32 +1,0 @@
-# @templ/esbuild-plugin-templ
-
-TEMPL esbuild plugin module.
-
-## `badges`
-
-<!-- automd:badges license provider=shields -->
-
-[![npm version](https://img.shields.io/npm/v/@templ/esbuild-plugin-templ)](https://npmjs.com/package/@templ/esbuild-plugin-templ)
-[![npm downloads](https://img.shields.io/npm/dm/@templ/esbuild-plugin-templ)](https://npmjs.com/package/@templ/esbuild-plugin-templ)
-[![license](https://img.shields.io/github/license/rjoydip/templ)](https://github.com/rjoydip/templ/blob/main/LICENSE)
-
-<!-- /automd -->
-
-## `pm-x`
-
-<!-- automd:pm-x args=. -->
-
-```sh
-# npm
-npx @templ/esbuild-plugin-templ@0.0.1 .
-
-# pnpm
-pnpm dlx @templ/esbuild-plugin-templ@0.0.1 .
-
-# bun
-bunx @templ/esbuild-plugin-templ@0.0.1 .
-```
-
-<!-- /automd -->
-
-[GitHub](https://github.com/rjoydip/templ/tree/main/plugins/esbuild-plugin-templ)
