--- plugins/esbuild-plugin-templ/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "@templ/tsconfig/base.json",
-  "compilerOptions": {
-    "baseUrl": ".",
-    "outDir": "dist"
-  },
-  "include": ["src/**/*.ts", "*.config.ts", "*.d.ts"]
-}
